+++ shared/src/commonMain/kotlin/dev/jketterer/leaflog/domain/models/Tea.kt
@@ -13,7 +13,7 @@ data class Tea(
     val purchaseDate: LocalDate? = null,
     val purchasePrice: Double? = null,
     val stockAmount: Int? = null,
-    val defaultTemperatureCelsius: Int? = null,
+    val defaultTemperatureCelsius: Double? = null,
     val description: String? = null,
     val photos: List<String> = emptyList(),
     val isFavorite: Boolean = false,
